fix: genre and rating id type migration to long

+++ src/main/java/ru/yandex/practicum/filmorate/controller/GenreController.java
@@ -22,7 +22,7 @@ public class GenreController {
     }
 
     @GetMapping("/{id}")
-    public Genre getGenreById(@PathVariable int id) {
+    public Genre getGenreById(@PathVariable long id) {
         return service.getGenreById(id);
     }
 }
